move reporting functions to new file
