+++ setup.py
@@ -1,5 +1,5 @@
 from setuptools import setup, find_packages
-from src import __version__
+from mmmvi import __version__
 
 setup(
     name="mmmvi",
@@ -8,5 +8,5 @@ setup(
     install_requires=["pandas", "pysam"],
     author="Dillon O.R. Barker",
     author_email="dillon.barker@canada.ca",
-    entry_points={"console_scripts": ["mmmvi=src.mmmvi:main"]},
+    entry_points={"console_scripts": ["mmmvi=mmmvi:main"]},
 )
